Ore: avoid tapping the same position multiple times
this doesn't help much - we probably need to rethink this algorithm
but it does help reduce the workload when generating large dirt seams.

It appears the intent of this algorithm is to try to generate overlapping
seams by slightly offsetting the base coordinates each time. This is OK for
small ore seams, but for large dirt seams, this causes it to attempt over 80%
of future placements in the same places we already placed in previous iterations.
This obviously doesn't make any sense, and we can avoid this with a hash
map to remember which positions we already tried.

In the future, we'll probably want to rewrite this algorithm to be less
insanely inefficient, but I don't fully understand it for now and don't
want to mess with it too much.