VanillaCommand: support @s command selector
this one is very easy to support since it always means "self",
so we don't need any fancy infrastructure to support it.
This should be way less annoying than remembering what the first
letter of my username is when having multiple accounts...

Other command selectors are generally much more complex, so
I won't bother with those for now. In addition, some of them
may select multiple targets, which our system isn't equipped for.