Separate sneak keydown flag from crouching flag (#6867)
This introduces a "sneak pressed" property, which may differ from "is sneaking".

Notably, the player may get forced to sneak when stuck under a slab in a 1.5 block high space even when they release the sneak key. In this case they should be able to interact with blocks as if they were standing tall.

Additionally, pressing sneak & space together while flying disables interacting with blocks without causing the player to physically sneak.

We probably should check the gameplay conditions for physical sneaking on the server side at some point, but this is good enough for now.

PlayerToggleSneakEvent will be called when the player presses or releases the sneak key, however if the resulting sneak state doesn't match what the player wanted (e.g. releasing sneak while stuck under a slab), the event will be cancelled by default. Plugins can use @handleCancelled to detect this.

closes #6066
closes #6733

Co-authored-by: Dylan T. <dktapps@pmmp.io>